feat : add livedata in one viewmodel
