+++ app/src/main/java/campus/tech/kakao/map/MainActivity.kt
@@ -20,7 +20,6 @@ class MainActivity : AppCompatActivity() {
         setContentView(R.layout.activity_main)
 
         val mapItemViewModel = MapItemViewModel(this)
-        val selectItemViewModel = SelectItemViewModel(this)
 
         val mapList = findViewById<RecyclerView>(R.id.mapList)
         val selectList = findViewById<RecyclerView>(R.id.selectList)
@@ -28,13 +27,10 @@ class MainActivity : AppCompatActivity() {
         val mainText = findViewById<TextView>(R.id.main_text)
         val cancelBtn = findViewById<ImageView>(R.id.cancelBtn)
 
-        //mapItemViewModel.updateMapItemList()
-        //selectItemViewModel.updateSelectItemList()
-
         val mapListAdapter =
-            MapListAdapter(mapItemViewModel.getMapItemList(), LayoutInflater.from(this))
+            MapListAdapter(listOf<MapItem>(), LayoutInflater.from(this))
         val selectListAdapter =
-            SelectListAdapter(selectItemViewModel.getSelectItemList(), LayoutInflater.from(this))
+            SelectListAdapter(listOf<MapItem>(), LayoutInflater.from(this))
 
         mapList.adapter = mapListAdapter
         mapList.layoutManager = LinearLayoutManager(this, LinearLayoutManager.VERTICAL, false)
@@ -45,13 +41,12 @@ class MainActivity : AppCompatActivity() {
         mapListAdapter.setItemClickListener(object : MapListAdapter.ItemClickListener {
             override fun onClick(v: View, position: Int) {
                 val mapItem = mapListAdapter.mapItemList.get(position)
-                selectItemViewModel.insertSelectItem(
+                mapItemViewModel.insertSelectItem(
                     mapItem.name,
                     mapItem.address,
                     mapItem.category,
                     mapItem.id
                 )
-                selectListAdapter.updateMapItemList(selectItemViewModel.getSelectItemList())
             }
         })
 
@@ -59,18 +54,24 @@ class MainActivity : AppCompatActivity() {
             SelectListAdapter.CancelBtnClickListener {
             override fun onClick(v: View, position: Int) {
                 val selectItem = selectListAdapter.selectItemList.get(position)
-                selectItemViewModel.deleteSelectItem(selectItem.id)
-                selectListAdapter.updateMapItemList(selectItemViewModel.getSelectItemList())
+                mapItemViewModel.deleteSelectItem(selectItem.id)
             }
         })
+
+        mapItemViewModel.mapItemList.observe(this) {
+            mapListAdapter.updateMapItemList(it)
+        }
+
+        mapItemViewModel.selectItemList.observe(this) {
+            selectListAdapter.updateMapItemList(it)
+        }
 
         inputSpace.addTextChangedListener(object : TextWatcher {
             override fun beforeTextChanged(s: CharSequence?, start: Int, count: Int, after: Int) {}
 
             override fun onTextChanged(s: CharSequence?, start: Int, before: Int, count: Int) {
-                val mapItemList = mapItemViewModel.searchMapItem(s.toString())
-                mapListAdapter.updateMapItemList(mapItemList)
-                if (mapItemList.size == 0) {
+                mapItemViewModel.searchMapItem(s.toString())
+                if (mapItemViewModel.mapItemList.value == null) {
                     mainText.visibility = View.VISIBLE
                 } else {
                     mainText.visibility = View.INVISIBLE
