+++ app/src/main/java/campus/tech/kakao/map/MapItem.kt
@@ -23,6 +23,15 @@ object MapItemDB : BaseColumns {
     const val TABLE_COLUMN_CATEGORY = "category"
 }
 
+object SelectItemDB : BaseColumns {
+    const val TABLE_NAME = "selectItem"
+    const val TABLE_COLUMN_ID = "id"
+    const val TABLE_COLUMN_NAME = "name"
+    const val TABLE_COLUMN_ADDRESS = "address"
+    const val TABLE_COLUMN_CATEGORY = "category"
+    const val TABLE_COLUMN_MAP_ITEM_ID = "mapItemID"
+}
+
 class MapItemDbHelper(context: Context) : SQLiteOpenHelper(context, "mapItem.db", null, 1) {
     val wDb = writableDatabase
     val rDb = readableDatabase
@@ -36,12 +45,24 @@ class MapItemDbHelper(context: Context) : SQLiteOpenHelper(context, "mapItem.db"
                     "${MapItemDB.TABLE_COLUMN_CATEGORY} varchar(10)" +
                     ");"
         )
+        db?.execSQL(
+            "CREATE TABLE ${SelectItemDB.TABLE_NAME} (" +
+                    "${SelectItemDB.TABLE_COLUMN_ID} Integer primary key autoincrement," +
+                    "${SelectItemDB.TABLE_COLUMN_NAME} varchar(15) not null," +
+                    "${SelectItemDB.TABLE_COLUMN_ADDRESS} varchar(30) not null," +
+                    "${SelectItemDB.TABLE_COLUMN_CATEGORY} varchar(10) not null," +
+                    "${SelectItemDB.TABLE_COLUMN_MAP_ITEM_ID} Integer not null," +
+                    "FOREIGN KEY(${SelectItemDB.TABLE_COLUMN_MAP_ITEM_ID})" +
+                    " REFERENCES ${MapItemDB.TABLE_NAME}(${MapItemDB.TABLE_COLUMN_ID})" +
+                    ");"
+        )
         insertMapItem(db, "카페", "서울 성동구 성수동", "카페")
         insertMapItem(db, "약국", "서울 강남구 대치동", "약국")
     }
 
     override fun onUpgrade(db: SQLiteDatabase?, oldVersion: Int, newVersion: Int) {
         db?.execSQL("DROP TABLE IF EXISTS ${MapItemDB.TABLE_NAME}")
+        db?.execSQL("DROP TABLE IF EXISTS ${SelectItemDB.TABLE_NAME}")
         onCreate(db)
     }
 
@@ -61,36 +82,6 @@ class MapItemDbHelper(context: Context) : SQLiteOpenHelper(context, "mapItem.db"
                 db.insert(MapItemDB.TABLE_NAME, null, values)
             }
         }
-    }
-
-    fun makeAllMapItemList(): MutableList<MapItem> {
-        val cursor = rDb.rawQuery("Select * from ${MapItemDB.TABLE_NAME}", null)
-        val mapItemList = mutableListOf<MapItem>()
-        while (cursor.moveToNext()) {
-            mapItemList.add(
-                MapItem(
-                    cursor.getInt(cursor.getColumnIndexOrThrow(MapItemDB.TABLE_COLUMN_ID)),
-                    cursor.getString(cursor.getColumnIndexOrThrow(MapItemDB.TABLE_COLUMN_NAME)),
-                    cursor.getString(cursor.getColumnIndexOrThrow(MapItemDB.TABLE_COLUMN_ADDRESS)),
-                    cursor.getString(cursor.getColumnIndexOrThrow(MapItemDB.TABLE_COLUMN_CATEGORY))
-                )
-            )
-        }
-        cursor.close()
-
-        return mapItemList
-    }
-
-    fun printAllMapItemList() {
-        val cursor = rDb.rawQuery("Select * from ${MapItemDB.TABLE_NAME}", null)
-        while (cursor.moveToNext()) {
-            Log.d(
-                "uin",
-                "" + cursor.getInt(cursor.getColumnIndexOrThrow(MapItemDB.TABLE_COLUMN_ID)) +
-                        "" + cursor.getString(cursor.getColumnIndexOrThrow(MapItemDB.TABLE_COLUMN_NAME))
-            )
-        }
-        cursor.close()
     }
 
     fun searchMapItem(category: String): MutableList<MapItem> {
@@ -120,5 +111,58 @@ class MapItemDbHelper(context: Context) : SQLiteOpenHelper(context, "mapItem.db"
         cursor.close()
 
         return mapItemList
+    }
+
+    fun insertSelectItem(name: String, address: String, category: String, id: Int) {
+        val values = ContentValues()
+        values.put(SelectItemDB.TABLE_COLUMN_NAME, name)
+        values.put(SelectItemDB.TABLE_COLUMN_ADDRESS, address)
+        values.put(SelectItemDB.TABLE_COLUMN_CATEGORY, category)
+        values.put(SelectItemDB.TABLE_COLUMN_MAP_ITEM_ID, id)
+
+        wDb.insert(SelectItemDB.TABLE_NAME, null, values)
+    }
+
+    fun deleteSelectItem(id: Int) {
+        wDb.delete(
+            SelectItemDB.TABLE_NAME,
+            "${SelectItemDB.TABLE_COLUMN_MAP_ITEM_ID}=?",
+            arrayOf(id.toString())
+        )
+    }
+
+    fun makeAllSelectItemList(): MutableList<MapItem> {
+        val cursor = rDb.rawQuery(
+            "Select * from ${SelectItemDB.TABLE_NAME} order by ${SelectItemDB.TABLE_COLUMN_ID} desc",
+            null
+        )
+        val selectItemList = mutableListOf<MapItem>()
+        while (cursor.moveToNext()) {
+            selectItemList.add(
+                MapItem(
+                    cursor.getInt(cursor.getColumnIndexOrThrow(SelectItemDB.TABLE_COLUMN_MAP_ITEM_ID)),
+                    cursor.getString(cursor.getColumnIndexOrThrow(SelectItemDB.TABLE_COLUMN_NAME)),
+                    cursor.getString(cursor.getColumnIndexOrThrow(SelectItemDB.TABLE_COLUMN_ADDRESS)),
+                    cursor.getString(cursor.getColumnIndexOrThrow(SelectItemDB.TABLE_COLUMN_CATEGORY))
+                )
+            )
+        }
+        cursor.close()
+
+        return selectItemList
+    }
+
+    fun checkItemInDB(id : Int) : Boolean {
+        val cursor = rDb.rawQuery(
+            "Select * from ${SelectItemDB.TABLE_NAME} where ${SelectItemDB.TABLE_COLUMN_MAP_ITEM_ID} = ?",
+            arrayOf(id.toString())
+        )
+        if(cursor.getCount() > 0) {
+            cursor.close()
+            return true
+        } else {
+            cursor.close()
+            return false
+        }
     }
 }
