+++ app/src/main/java/campus/tech/kakao/map/SelectListAdapter.kt
@@ -8,7 +8,7 @@ import android.widget.TextView
 import androidx.recyclerview.widget.RecyclerView
 
 class SelectListAdapter(
-    var selectItemList: MutableList<MapItem>, val layoutInflater: LayoutInflater
+    var selectItemList: List<MapItem>, val layoutInflater: LayoutInflater
 ) : RecyclerView.Adapter<SelectListAdapter.ViewHolder>() {
     inner class ViewHolder(itemView: View) : RecyclerView.ViewHolder(itemView) {
         val name: TextView
@@ -47,7 +47,7 @@ class SelectListAdapter(
         return selectItemList.size
     }
 
-    fun updateMapItemList(mapItemList: MutableList<MapItem>) {
+    fun updateMapItemList(mapItemList: List<MapItem>) {
         this.selectItemList = mapItemList
         notifyDataSetChanged()
     }
